Extract the compiler's Frame into its own file
